+++ CHANGELOG.md
@@ -30,7 +30,9 @@
 
 ## [Unreleased][]
 ### Added
+- Readme: readme is shown as popup on first installation and available at the settings menu.
 ### Changed
+### Removed
 ### Fixed
 
 ## [2.6.0] - 2024-10-09
@@ -42,6 +44,7 @@
 - [JSplitter (SMP)](https://foobar2000.ru/forum/viewtopic.php?t=6378&start=360) support and ES2021 compatibility.
 - Helpers: in case saving a file throws an error due to long paths (+255 chars) a warning popup will be shown.
 - Helpers: updated helpers.
+### Removed
 ### Fixed
 - UI: compressed files don't display neither an "incompatible format" nor the visualizer as fallback.
 - UI: 'Visualizer during analysis' setting was overriding the 'Visualizer for incompatible files' in some cases, displaying the visualizer indefinitely for incompatible files (and not just during the analysis).